Change remove to require distinct if used.
In our use case (contacts) we want to be able to remove names if it changes, but just because we call `trie.remove('mark')` doesn't mean we want to get rid of all the marks we know. Just a specific distinct one - adding the distinct parameter makes it easier for us to do so.